# RentalManager.WebApp/src/references/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RentalManager.WebAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/RentalEquipment/{id}": {
      "delete": {
        "tags": [
          "RentalEquipment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "RentalEquipment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RentalEquipmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalEquipmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalEquipmentDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RentalEquipment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRentalEquipment"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRentalEquipment"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRentalEquipment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RentalEquipmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalEquipmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalEquipmentDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "RentalEquipmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateRentalEquipment": {
        "required": [
          "name",
          "price"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "pattern": "^[a-zA-ZąćęłńóśźżĄĘŁŃÓŚŹŻ 0-9\\/.\\-\\+]*$",
            "type": "string",
            "default": "Fun equipment mk III"
          },
          "price": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "default": 100
          }
        },
        "additionalProperties": false
      }
    }
  }
}